refactor: 重构代码并优化依赖管理
- 添加 Hutool 工具库依赖
- 重构 RoleDto 类，简化代码结构
- 使用 Hutool 工具类替换 CommonUtils 中的集合操作
- 修改多个控制器的 HTTP 方法从 PUT/DELETE 到 POST
- 清理配置文件中的注释代码
- 优化角色资源分配逻辑
- 重构 CommonUtils 工具类，使用 Hutool 实现

+++ module-user/src/main/java/exam/demo/moduleuser/pojo/vo/RoleItemVo.java
@@ -56,7 +56,7 @@ public class RoleItemVo extends BaseVo implements Serializable {
     /**
      * 状态位
      */
-    private Byte status;
+    private Long status;
 
     /**
      * 资源节点ID
